Refactor quantity handling across multiple models and views
- Changed quantity type from REAL to INTEGER in invoice_products_table and order_products_table.
- Updated InvoiceProduct and OrderProduct models to use int for quantity instead of double.
- Adjusted mock data generation to produce integer quantities.
- Modified business snapshot report to calculate quantity sold using conversion factor.
- Updated OrderListBloc to handle quantity updates with conversion factor.
- Refactored form product model to use int for quantity.
- Changed quantity parsing in create_invoice_page and edit_invoice_page to use int.
- Updated latest_sales and top_product_activity views to display customer names correctly.
- Adjusted quantity handling in quantity_and_unit.dart to use int.
- Updated print_receipt.dart to format quantity as int.
- Changed data_row_mapper.dart to display "Walk-in Customer" for unnamed customers.

+++ lib/domain/models/invoice_product.dart
@@ -24,7 +24,7 @@ class InvoiceProduct {
         productId: map[InvoiceProductsTable.INVOICE_PRODUCTS_PRODUCT] as int,
         productName: map[InvoiceProductsTable.INVOICE_PRODUCTS_NAME] as String? ?? '',
         description: map[InvoiceProductsTable.INVOICE_PRODUCTS_DESCRIPTION] as String?,
-        quantity: (map[InvoiceProductsTable.INVOICE_PRODUCTS_QUANTITY] as num).toDouble(),
+        quantity: (map[InvoiceProductsTable.INVOICE_PRODUCTS_QUANTITY] as num).toInt(),
         secondaryUnit: map[InvoiceProductsTable.INVOICE_PRODUCTS_SECONDARY_UNIT] as int?,
         conversionFactor: map[InvoiceProductsTable.INVOICE_PRODUCTS_CONVERSION_FACTOR] != null
             ? (map[InvoiceProductsTable.INVOICE_PRODUCTS_CONVERSION_FACTOR] as num).toDouble()
@@ -42,7 +42,7 @@ class InvoiceProduct {
   final int productId;
   final String productName;
   final String? description;
-  final double quantity;
+  final int quantity;
   final int? secondaryUnit;
   final double? conversionFactor;
   final double rate;
@@ -54,7 +54,7 @@ class InvoiceProduct {
     int productId,
     String productName,
     String? description,
-    double quantity,
+    int quantity,
     int? secondaryUnit,
     double? conversionFactor,
     double rate,
